stringify content, remove HcaUtils
